requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
pillow==10.1.0
fastapi-limiter==0.1.5
redis==5.0.1 